requirements.txt
crewai==0.51.1
crewai-tools==0.8.3
langchain-google-genai==1.0.8
python-dotenv==1.0.1